// src/shared/api/subsonic/subsonic-schema.d.ts
export interface paths {
// ...
    "/rest/getAlbum": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Returns details for an album.
         * @description Returns details for an album, including a list of songs. This method organizes music according to ID3 tags.
         */
        get: operations["getAlbum"];
        put?: never;
        /**
         * Returns details for an album.
         * @description Returns details for an album, including a list of songs. This method organizes music according to ID3 tags.
         *
         *     Requires OpenSubsonic extension name `formPost` (As returned by `getOpenSubsonicExtensions`)
         */
        post: operations["postGetAlbum"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}